requirements.txt
Flask==3.0.2
lxml==5.1.0
Pillow==9.0.1
Pillow==10.2.0
protobuf==3.12.4
Requests==2.31.0
Werkzeug==3.0.1
